+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2026-01-26
+
+### Added
+
+- Force folder sidebar view for non-admin users in Media Library
+- Non-admins can no longer toggle to list/grid view without folders
+- New `MediaLibraryEnforcer` component with separate JS/CSS assets
+
+### Changed
+
+- Refactored to use external asset files instead of inline scripts/styles
+
 ## [1.3.5] - 2026-01-25
 
 ### Changed